Main/Editor JSX!
Heck yeah! This actually worked out really nicely. The props works great now too, or rather the global state/signals. It seems to be much easier to build this from the bottom up, rather than from the top down. Seems like a similar, more easy to go about method, like how you migrate from JS to TS: going from modules with no dependencies, up to ones that require those ones. Then you fill in the gaps for things that aren't properly connected yet. Laying the foundation for your adjustments to fix later seems to work much nicer than trying to plan things out and work downwards. Absolutely going to keep this in mind for React in general, and migrating other projects over to this component way of doing things. So happy things are working nicely! Wanting to improve on this for some time now, and I really like how SolidJS is clicking for me.

The concept of opting-in to reactivity seems so much simpler to me, like in terms of a mental model. Having that in mind, it makes sense why React is so useful, because everything essentially just works by default. You have to work the other way to make things performant again. Instead with this kind of setup here, you have to work to getting to overly complex things, rather than having to work towards simple/performant ones. Both have their merits, and one may be better for a given project than the other. All up to interpretation and a point of reference :)

https://www.solidjs.com/docs/latest/api#createeffect
https://developer.mozilla.org/en-US/docs/Web/HTML/Element/textarea#autocorrect