Tree View Value Debug
Was looking to commit this the other night, but I was getting too tired to write the commit description haha.

I discovered that it is indeed the render function for the `<NBTBranch/>` component that's preventing the updates from happening correctly. When changing what it renders to something statically analyzable (from what it seems), Solid is able to detect the change of `getTreeViewValue()`. So the `switch (getType())` wrapper does seem to be at fault I think. I realized I can add a temporary `createEffect()` at the top of the component which will toggle it's visibility when the state changes, allowing it to rerender when the `getTreeValue()` state updates. That's more of a hack though, I want to get it working properly eventually.

https://www.solidjs.com/docs/latest#createsignal

Forgot to mention in my music discussion the other night, Shoegaze is now another genre I think I may look into. It's really intriguing me, and it sounds like something I haven't looked into yet, unlike sometimes when you look into subgenres and it sounds like other kinds of things you already know. This one sounds fresh. I started trying to make a Shoegaze song last night, but it needs more effects (lol), and I'm debating how the drums should sound still. I have a few different concepts for the beat, but I don't know if it should be a classic drum sound, or be a compressed electronic one.