NBTBranch Proper Control Flow
Yay, this fixes it! It properly knows how to traverse the reactive parts of the JSX. My overarching `switch (getType())` statement didn't allow it to follow the path of what should change.

https://www.solidjs.com/tutorial/flow_switch (This is another alternative I may be able to use. I wanted to see if I could built it using JSX/language primitives, before needing to use a Solid-specific feature. So happy that Solid allows you to use one or the other!!)

Also realized the `treeViewToggle` element's checked value was bound to the wrong state accessor, oops!

Seeing the JSX for this more simply laid out in a single group is helping me see what the most effective way might be to separate the recursive component out into it's own one. That would mean that one could be for value-based tags, and one could be for container-based tags, where it renders children of other values. I do like the ability to only need a single component to wrap these though, so maybe I will keep the recursive setup, so then you don't have to call a specific component based on what kind of tag you are putting in. Maybe it could have an overarching wrapper to handle that for you though, so each tag can get it's own component, but you only call it with a single generic one. I like the idea of that. That may work really nicely in conjunction with the Solid `<Switch>` component as well!