PrimitiveView / ContainerView
Trying to better distinguish between what kinds of tags are what. I tried implementing a custom component for each of the tags on their own, but it ended up being essentially exactly the same, so I'm not going to do that for the time being. If the HTML for the Tree View does change (eventually will), then I will look into having individual components for each tag. Right now the only thing to render differently is the `data-type` icon next to the tag entry, and this is currently still scoped by `NBTBranch`. I would like to figure out how to move that into the component itself maybe, rather than needing to use each of these on their own. But, we'll see! By that I mean needing to use both a wrapper `NBTBranch` in conjunction with the underlying components. It seems much simpler now, but I think I can even abstract other parts of it out.

I am really liking the accessor model that Solid provides. It very much helps me easily keep track of what is reactive, versus what is not. There have already been a few pitfalls on my behalf of nearly using static values instead, but then I remembered that those won't properly re-render.