Format Options Binding (Experimenting)
My current solution here isn't the right solution, but now I better understand how to use the proper solution. The concept is that you can instead use a piece of state to manage the `<form>` elements' values, and just use that to maintain what the current value of the options are. Seeing the concept for that now, I really do like that! It is great that the browser has vanilla support for `RadioNodeList` too though, that works maybe even more magically than the state setup does.

https://legacy.reactjs.org/docs/forms.html
https://stackoverflow.com/questions/70019647/how-to-handle-group-radio-button-in-react-js-functional-component