Format Options State Binding!
Yay, this is fully working, and it's great! I think that means I can kind of dissolve the use of the `openOptions()` and `saveOptions()` functions, at least by abstracting them into the only place they are used, which are the `openFile()` and `saveFile()` options.

I think I was nearly to this kind of point when trying things out in Svelte, but this is working even more nicely than that! The only downside I see is that there isn't a 1:1 way of using `RadioNodeList`, which is a nice way of managing the state for things, without needing to use events. It's not completely simply to use for non value-based state though, kind of like how I had to do the custom handling for the `BedrockLevel`, `RootName` or `Compression` types, for example. `null` isn't a specific form value for those entries, but rather means an empty `<input>` value, or a toggled checkbox depending on the given concept.

https://developer.mozilla.org/en-US/docs/Web/API/HTMLInputElement#instance_properties (valueAsNumber is great!)
https://davidwalsh.name/input-valueasnumber
https://caniuse.com/mdn-api_htmlinputelement_valueasnumber (I was fairly sure it wasn't a new/non-standard feature, but I thought I'd double check before I got too excited)