FormatOptions Component
Working on moving each of the component stacks to their own modules, but I need to figure out how to properly manage state from the top down first. Right now there is still a global state thing, where `App` uses state defined outside of itself, and `FormatOptions` needs to access it as well. I need to properly move every component to it's own module, and out of `app.tsx`, because moving only `FormatOptions` to it's own file causes recursive module imports, which isn't doable from this current tooling (while ESM does allow for it, it's not the best thing to rely on since other tools don't quite expect to handle it properly. You can do it in vanilla ESM, but I think it's a compiler issue maybe)

Listening to "Boil That Dust Speck" some more! I feel like I've heard "Skunk" sometime before, maybe that live Beer for Dolphins video from the late 90s, it was so great! It's on YouTube, I think there's 3 videos or so.