Element Listener Organization
Natty Trousers & Scotch
Scotch sounds like Alice in Chains kind of to me, like a "grunge Keneally" sound. Heck yeah!

Was going to try function binding the use of `setShowTreeView()` so I didn't have to use the `() => ` wrapper call, but I think it unnecessarily complicates the look of things, and it's not obvious as to why I really need to do that. Same with doing `onclick={fileOpener.click}`. In concept it works, but `fileOpener` has to be defined beforehand, meaning that assigning things in this way doesn't truly work, because of the JSX structure. Back when I was using element IDs, the elements are already exposed as variables on the global scope, by the time the JS runs, so that's why you could access them beforehand.

Realized the `onchange` handler for the disableName input wasn't fully set up yet.

https://stackoverflow.com/questions/1986896/what-is-the-difference-between-call-and-apply
https://www.keneally.com/music/discography-mike-keneally/boil-that-dust-speck/