Alien + Side-Effects
Checking out the full Strapping Young Lad Alien. I'm loving it. Love?

Trying to better scope the logic for various parts of the app's handling, so things are either stateful, or they aren't. I tend to have made things in the past which are fairly side-effect driven, making it hard to use on their own individually. I want to try and counteract that a bit.