Open File Elision
Trying to further flatten the use of things in the functional stack. Listening to the Devin Townsend Podcast for Alien at the moment, so that's why it's hard to write this.

Might not use the FS Access API for the file opening process, afterall. It limits the use of files that you can open, in that it can't be system-like files (so no `.minecraft` directory), and turns out you can't open `.dat_old` files either. I think if you want to be able to save directly back to a file, you can still do so by dragging it in drag and drop. Otherwise, it will let you open any file you'd like, by use of the `<input type="file">`. So either saving directly can work, or you can open a broader set of files. Maybe I can provide an option for this in the GUI when you open a file actually, could be a simple way to handle that.

https://developer.mozilla.org/en-US/docs/Web/API/Window/showOpenFilePicker
https://stackoverflow.com/questions/76970742/how-to-read-a-file-using-showopenfilepicker-not-input-type-file
https://developer.chrome.com/docs/capabilities/web-apis/file-system-access
https://wicg.github.io/file-system-access/#dom-window-showopenfilepicker
https://wicg.github.io/file-system-access/#wellknowndirectory-too-sensitive-or-dangerous