Component Prop Scoping: Header
Going to learn how to move state for each component into being props instead. I haven't liked this pattern because needing to pipe things as props has seemed counterintuitive to me thus far. I think if you build your component trees nicely though, then it shouldn't be an issue. So maybe it's less about the need to pass things through each component as props, and more that I possible wasn't managing the component state properly before. Say like for the `showFormatDialog()` prop, previously I would pass in `formatDialog` so you could call `formatDialog.showModal()`. But now I think the concept is that it should instead take a callback, which contains that call instead. So you don't want to have to pass in the dependencies of the scope for that event handler, but rather instead pass in that function itself, because you can make it in a scope/context from where those are lexically accessible, meaning you don't have to worry about what that function calls, when passing it down to children components like Header. I think this is a bigger way to look at things which I previously didn't understand, or see before!

I would like there to be a way to pass a plain Signal to components as well, rather than passing in both the Accessor and Setter separately. Is there a nice way for doing that? I kind of wish that the signals would return something like `isDisabled.get() / isDisabled.set()`, rather than having to destructure and name them separately. If you don't destructure the raw Signal type from the `createSignal()` call, that makes them an array, and you have to use indexing to index each one, which I'm not too fond of. Essentially, `isDisabled[0]()`.

While I do really like using global state management (it seems to come the easiest to me), I definitely see now how it makes working with components on their own more difficult to set up. It makes it harder to do instancing with things, because there wasn't already instancing built into the higher component's state, meaning you can't easily duplicate that higher element if you wanted to. This also goes with the issue of global state implicitly tying a child component to that of it's parent component. Seeing it like that now, I can see why my previous setup wasn't translating well into a component structure, because it wasn't setup properly where each component could work interchangeably given any context it was in. You needed to have the parent's state in order for the child to work. It's one thing to have required props to make the child work correctly (this new concept here), but if you don't allow for an explicit IO of what the component needs, and allow those hooks to be supplied from anywhere you'd like, then you can't actually detach it from the parent's context. It's like a function with side-effects. You can't move it to it's own module if it's accessing variables from it's parent scope. You want to move those over to parameters instead, because you can supply those from any context, it doesn't matter where the function is defined or called. I'm not sure I completely conceptualized things in this same manner as with components before. Or if anything, I didn't see why that would be useful, if you just use the component once.

I think this possibly could have a similar trade-off for classes as well. There have been a few cases where methods on my own classes would be useful to have available as standalone functions outside of the class's state. So I'm curious about the trade-off between using class-based state of a given operation, compared to parameter-based dependency tracking. Similar to the concept here, I have been worried about needing to pass too many parameters/state around with each call, and that using a class seemed more simple to me. It inherently couples the feature to where it's state is defined, which means you can either only use it in conjunction with that state instance, or by emulating the environment of that instance.

This actually goes very similar to that of operating system virtualization and whether a program is cross-platform or not. If a given concept is intrinsically tied to it's underlying implementation, it isn't easily migratable to a new host location. Say like an x86 binary. It works just fine on an x86_64 machine, because it is compatible there. It won't work on an ARM64 machine without emulation though, because the low-level calls are inherently tied to that of the x86 platform. If you have the higher-level (technically speaking) code for that program though, you could re-compile it's same logic to native platform calls, because the programming language is what bridges the gap of what is different between the two system architectures. So in that scenario, using C instead of x86 Assembly is like using parameters instead of global state.

I think this is why I have liked using JS for things as well. It may not be the most efficient language, but has very low coupling to the platforms that you run it on. You don't need to compile it to use it on a platform different than to what you wrote it on. Same with this new concept (to me) for UI components. If you use parameters instead of global state, you can run them in any context or in any parent container, unlike where with using global state, the component is intrinsically tied to that of it's parent.

FP, this is the way 🗣️