State Input Syncing
Realized that I was using the `onchange` events, instead of `oninput`, which was what behavior I wanted instead for the FormatOptions menu.

Also discovered that I forgot to update the value of `setEditorValue()` from the SNBT `<textarea>`. I hadn't yet ensured to debug the output that `saveFile()` was making with the new UI stack in JSX.

https://stackoverflow.com/questions/38256332/in-react-whats-the-difference-between-onchange-and-oninput (I know I'm not using React, but still helpful to learn)
https://stackoverflow.com/questions/17047497/difference-between-change-and-input-event-for-an-input-element

I think `onchange` was fine for checkbox and radio inputs, but I forgot to think about the cases for text and textarea inputs.

Discovered the other day that it's a very awesome crossover from the beginning of Empath "Sprite", back to the "City" demos. It's the same tale about the bird without a mom or a dad.

OOOH! I just realized even more. The first store was told by a young child, and the Sprite version is by an old man it sounds like. I didn't think anything of that, but I wonder if it's a detail in that a lot of time has passed, and the young child has grown into being an old man, and is still telling this story this many years later. Wow that's a great detail, if that's what that is :)

Getting to see Devin play live tomorrow!