Cross-Component Element Referencing
Heck yeah! This has been a challenge for me to figure out too. You can use state to manage this turns out, make sense.

Hadn't realized yet that Elon is in "Singularity", haha. Makes sense.

https://stackoverflow.com/questions/37647061/how-do-i-access-refs-of-a-child-component-in-the-parent-component

Was trying to migrate things to individual modules again, and this was another holdup I hadn't remembered to account for just yet.

This model of exposing props and bringing the data closer to the component entry points is very cool, I hadn't thought of it like this before. It's like providing electrical outlets to your kitchen, rather than needing to cut a hole in the wall, and tie your coffee maker to the powerline on the street. The other version is that you can use global state, and get an extension cord and connect your coffee maker to your powerline by going through the backyard instead. Both of them are patch fixes, the hole one being more messy and hacky, while the other one is a bit of a mess, but manageable. Having power outlets avoids needing to hack things together, because they each allow for ways to connect each other together. Modularize!