Stateless File Saving
Trying to generalize the use of various utility functions, so they don't need state to function. They should only require what data they need to do the action, and it should be a 1:1 IO operation. No side-effects! :D

The API for things, because of this "un-sideeffect"-ication, is fairly messy now, but that's okay because now there's a simply visible correlation between what data a given feature needs, rather than being hooked to a specific component.