Open/Save State Management
Realized part of my struggle in the previous setup! It was that I was trying to setup the generic file handling too close to the state of the app. The solution is to implement those simply with lower-level return statement-based functions, and have separate ones that call those under the hood, which instead deal with the state of the app outside of those IO ones. That's what I did here.

I have also been looking into how I should work with global event handling when using a component-based setup. Should they be declared unrelated to the `<App/>` component? I think that would be going against the point of them, but I'm not totally sure.

https://www.reddit.com/r/reactjs/comments/w8hdd5/how_to_handle_keyboard_events_the_proper_way_in/
https://www.reddit.com/r/reactjs/comments/qkap4d/react_rendering_on_keydown_event/
https://stackoverflow.com/questions/29069639/listen-to-keypress-for-document-in-reactjs
https://www.solidjs.com/tutorial/lifecycles_oncleanup
https://caniuse.com/mdn-api_filesystemfilehandle (Not sure if I should be using `instanceof FileSystemFileHandle` since that will error if you are using it on objects that aren't defined yet)

So I think the goal with these updated functions is that you can use/define the generic functions separately from the `<App/>` component, and the more `<App/>`-centric state management is scoped to that only, rather than being part of the generic functions themselves.