Generic File Functions
Related to the previous commit, now the universal file handling that isn't state-bound is not usable outside of the scope of any given component! :D

I hadn't realized this concept before either, whether a given function/feature should be tied to the state of the component or not. If you don't relate it to any state, and you only use the function's IO (parameters and return values), then it can be used anywhere you'd like, rather than only from a single component. This, is the way.

I think I will also look into moving the error handling for `readFile()` back over to the main module as well, since that's actually where it should be derived from.