Dedicated Service Worker Lib Types
Coming back to investigate how one can properly type the APIs for a Service Worker. I don't fully like the need for a second/separate folder for this, but hey it seems to work. I like being able to use `tsc` and it work for all of the source files on the project. That isn't possible here, since you'd have to run it twice for each TSConfig. I was wondering if project references could help with this, but I don't think it seems like it can. I might not be using it right though.

https://www.typescriptlang.org/docs/handbook/project-references.html
https://stackoverflow.com/questions/74786101/what-is-the-correct-tsconfig-json-for-a-service-worker
https://joshuatz.com/posts/2021/strongly-typed-service-workers/

Was wondering if I could add the hash to the Service Worker, and import that using the `?url` import, rather than plainly using `register("./service-worker.js")`.
https://vitejs.dev/guide/features