File Opening Error Handling
Didn't know about the DOMException class! It allows for exactly what I was looking for, once again. Had to debug it with `console.dir()`, because I think it just shows up as a simple string in the console.

The concept behind it seems to be that you can find out what kind of error occurred, without having to have different error objects on their own. This is exactly what I was looking for, and I was hoping I could find out what kind of thing went wrong with the file IO call, and handle each situation differently. Now I can! Aborting is different than another error. If the file picker is closed, the Promise still rejects, but the result is ignored. If it errors because the file couldn't be read, the user still gets and alert with the error message. That's the exact behavior I was looking for. I'm surprised this wasn't documented somewhere for the FS Access API pages, maybe I just missed it on there somewhere (likely, yeah).

I also reverted the previous demo for the file extension validation imports, I will work on a better, proper way to handle that a little more down the road.

Glad to have this new handling, it also allowed for fixing the bug around making the editor work again if the window is aborted, since while it opens the file, it disables the header and editor views. Now it can reenable it when the user returns if they didn't pick a file. I still need to handle this behavior a little better in some spots, like whether some inputs should keep their value, if the newly opening file fairly to be read for example. Same with the Format Options dialog. If you check Unnamed Root Tag, should the name in the input box be removed (current, unintentional behavior), or should the value be set to empty, so when it is unchecked, it defaults back to ""? I think the value should persist, since that seems the most GUI-friendly from a UX perspective. The value under the hood is indeed set to `null` when you check the unnamed option, but I feel like the GUI should persist the named version when only checking the box. When adding another file though, I think that would overwrite the input's value, since you don't want that to accidentally stick around to when you may uncheck the unnamed option, say if the new file were unnamed as well.

This is the kind of UI state stuff that I love figuring out. Having a UI framework for that is already so much more intuitive and inspiring/hopeful, because it's something that's much simpler for you to keep track of. Eventually moving STE over to this model is going to be outstanding!!!

https://developer.mozilla.org/en-US/docs/Web/API/DOMException
https://developer.mozilla.org/en-US/docs/Web/API/DOMException/name
https://developer.mozilla.org/en-US/docs/Web/API/DOMException#error_names