Unnamed Field Persistence (Partial, demo)
This doesn't fully work exactly just right, yet. Not completely sure the best way to handle this yet. It is very close though! Using a memo, I can keep track of the previous value. Right now, it does indeed persist the previous one in the input, and it gets updated with each change to the name. If the `getRootName()` value is null, the value will persist. This is a little of an issue though, in that like I mentioned before, opening an unnamed file afterwards doesn't properly set the content to "". So I want it to not update while editing the value yourself, but I want it to update when Dovetail updates the state (say, like for opening a new file).