Stateful Dialog Visibility
Ahah! Figured out another misconception of mine. Now I realize that the visibility state of the dialog is what you want to manage as a piece of state/prop, rather than the callback that invokes that. This makes things much simpler in terms of what the component tree looks like as well, since now `<App/>` doesn't need to know about the internal dialog for `<FormatOptions/>` anymore.

Now I'm starting to see how state is one thing, and how the app's reactions to those values change, is another thing.

https://stackoverflow.com/questions/76567184/html-dialog-element-in-react-js
https://developer.mozilla.org/en-US/docs/Web/API/HTMLDialogElement/close_event