Global Event Handling Functions
Working on setting up a proper cleanup step for the main `<App/>` component. I don't think `window.launchQueue.setConsumer()` can be reverted it seems like though? Not sure how I can clean that one up. I can remove the other events though, that should work nicely I think.

This is just the first step, I'm moving the event functions themselves into their own declarations, so I can remove them as needed later on during the cleanup step, say if `<App/>` were to be un-rendered. Is it commonplace to design components to work in tandem with one another, at least for ones as big of a scope like `<App/>`? I would like that to work, ideally, but maybe that only makes sense for non-"rootish" components, since you pretty much expect them to just be a singleton anyways. What are your thoughts on that? (Not sure who 'you' is in that context, who ever is reading this :P)

https://stackoverflow.com/questions/32896624/react-js-best-practice-regarding-listening-to-window-events-from-components
https://www.solidjs.com/tutorial/lifecycles_oncleanup?solved

Didn't realize just how much my sound comes from Math Rock, how much I like Math Rock, as well that it seems to be what Marco's style is as of the more recent years too, at least partially. I want to make some sort of mix between Math Rock and Metal, maybe it can be Math Metal? Sounds similar to other things out there that already exist, but the concept I hear it to be would be a bit different than regular Prog Metal. I'd like the personality behind Math Rock, it feels very close and personal, like being with someone you know. It's that part that I want to tie in to heaviness as well!
https://www.youtube.com/watch?v=BM-Mh-LLRu4
https://www.youtube.com/watch?v=UZujze-Q86Y
https://www.youtube.com/watch?v=1aZyiIXbWDc