Launch Queue Learning
Demo for learning how Launch Queue works with the Launch Handler manifest configuration. This works nicely with opening other files in the same window. I was looking to do this same thing with STE, since that supports editing multiple files at once. Once I implement a version of that for Dovetail as well, then I can do this feature here as well. This is a change just to demo things.

Also found my module naming was goofed from the file renaming in the previous few commits. Also, I still want to use `.js` extensions to make things more assuredly ESM out of the box, if possible. Say if I wanted to transpile the JSX to a plain ESM file, the import extensions would still be correct for running in the browser. Maybe I can configure TS to only allow `.js`? I know it prevents `.ts` ones by default, but it kept the original file name as `.jsx`.

https://developer.mozilla.org/en-US/docs/Web/API/LaunchQueue
https://developer.mozilla.org/en-US/docs/Web/Manifest/launch_handler
https://developer.chrome.com/docs/web-platform/launch-handler

Woah, this is a freaking cool find too!
https://developer.chrome.com/docs/web-platform/element-capture