Spacing Fixes (Temp)
Was diffing the current working version here, with the stable build in production, and realized the formatting/spacing of a bunch of small things was off. Turns out it was because of how I indented my original HTML! Funky thing. I think I encountered that while trying to move to React a little bit ago. This is a temp fix to come back to later as well, as I want it to simply look like the original for the time being, without any inconsistencies (even if that means replicating that original, unintended behavior). I plan to properly use Flex and Grid gap instead, once I move towards reworking the styles and markup shapes a bit more, now that things are more modularly written in JSX.

It was indeed the JSX that caused this to change over time, as the last commit to have all of these looking fine was when everything was still HTML-based, and the next commit was written in JSX, hence why the formatting was lost.

1c1de09 (last one to look like production, now it looks like this again)