Tree View Toggle - Form ARIA
Removing the IDs from all of the elements made it so some of the additional accessibility parts of the original markup was lost, so this strengthens that again. What's the best way to look into learning the best practices for ARIA roles and accessibility? I want to get started with that, and I don't want to make the wrong choices trying to be more specific, only to make things worse by adding incorrect roles, over just the implicit roles that various elements get.