Root Name Tree View Rendering
Super awesome! Now the tree view will render the name of the Root Tag, and it also updates live to when you edit it's value from inside of the Format Options dialog!

Learned the difference between the `<b>` and `<strong>` tags, as well as `<i>` and `<em>`. The second of the two are the semantic versions, meant to convey a difference in meaning to the text that surrounds it, while the single-character tags are meant only for appearance differences.

Realized a new way I can do stacked ternary organization. It allows for better examination of understanding the control flow of the statements/checks. Now it's the early-return kind of syntax, where it leaves the ternary when the condition is met, rather than being an exclusionary one. Essentially, you just invert the checks in each state, and it allows you to describe them in a line going downwards, without having to "nest if statements", or what the equivalent of that would be for ternary terminology.

Been listening to more of my latest songs I've been working on (after having not been too close with them intentionally, to let time pass for metal room to work on them), and I am realizing just how hard I am on myself for how things sound. They are very good, and I really like them. It kind of caught me by surprise how much potential there was with some of the demo songs I didn't keep for the full album. Learning to work on keeping my head up towards how I feel with my progression in things, and to remember it just may need a little breathing room to see the whole picture on how you are currently doing for things like your goals and what you want to accomplish.