NBTData Default
Now there is a default `NBTData` object that is used when a file hasn't been loaded yet. I think this goes in favor of better allowing files to be made arbitrarily as well, since this paves the way for allowing new files to be made from the beginning of opening the app.

#6