FS Access Picker Removal
Looking further into both `window.showOpenFilePicker()` and `window.showSaveFilePicker()`, neither can access files in "dangerous" or system-related folders, which rules out a good deal of places where NBT is stored. I don't want that to limit people in using the app, so I'm not going to support that. I thought maybe that you could open a file using a regular file input, then pick where you'd like to save it with `showSaveFilePicker()`, but looks like that won't let you save it there either. Thankfully you can still get nice and easy direct file saving if you had dragged the file in from your file manager. That works no matter the directory it's in, or what file it is. The two picker functions also don't work with `.dat_old` extensions, which is a bit annoying as well. So with those combined, I'm going to stick with drag and drop in combination with the File Handling API to allow users to open files from where they'd like. Using a regular file input from the app itself will work to open the file, but you don't get a file handle when using that itself. Same with saving the file at that point, it won't be able to remember where you want to save that file when you click "Save" from the GUI. This is a bit annoying/hectic for making new arbitrary NBT files as well, since you aren't able to drag the source of that in from the file system. So I may have to make a workaround for adding direct support somehow or another.