SNBT Editing Support!
This was actually not too bad to implement, heck yeah! Now that everything is modular and more-explicitly defined in it's own groups, it's easier to tell where something should be implemented, and it works across the whole rest of the app as well. Next I'd like to derive the `<input type="file">` picker from the Web App Manifest's file types as well, then it's also only defined from a single place too.

#6