Nearing Production + Input Cancelling
https://stackoverflow.com/questions/34855400/cancel-event-on-input-type-file
https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/cancel_event
https://developer.mozilla.org/en-US/docs/Web/API/HTMLInputElement/invalid_event (Don't think this applies to `<input type="file">`, or at least I couldn't find any info that it does. Is there anything that file inputs can error with?)

Weird freakin song, uncannily happy, kind of like if depression had a sound I think.
https://www.youtube.com/watch?v=xI0Htkrsc8Y

Working on bringing this branch up to par with running in production! Nearly everything is synonymous with the current build, so I want to start dogfeeding it to find any other issues, especially with the move to SolidJS.