MIME Types Dedupe (Test 2)
I think this is the second commit of trying this out. At least, I did try it before, but I may have not committed it.

The goal is to only define the accepted MIME types and file extensions in a single place, then the rest of the app can use them from that declaration as well. STE would greatly benefit from this as well.

https://github.com/tc39/proposal-json-modules (Looking forward to this, I'd like to use it with my Web App Manifest. Vite doesn't seem to support it either, unfortunately; it just treats the default import as a URL string, even if you omit `?url`.)
https://vitejs.dev/guide/features#json
https://github.com/WebAssembly/esm-integration
https://github.com/WebAssembly/esm-integration/tree/main/proposals/esm-integration