Web App Manifest Types (Concept)
Wanting to get full typings for the Web App Manifest, but for some reason it's not part of the TS standard library. There is an `@types` package on npm, but it's not complete with all of the fields that are part of the spec. May look into writing my own file, or maybe even a package for it as well.

Was curious if New-JavaScript had types for it as well, but looks like not, kind of surprised hehe. Maybe I can contribute to that instead, could make sense.

https://www.npmjs.com/package/@types/web-app-manifest
https://developer.mozilla.org/en-US/docs/Web/Manifest (This is what I'll use for the types)
https://developer.mozilla.org/en-US/docs/Web/Progressive_web_apps/Guides/Making_PWAs_installable#required_manifest_members (Not all fields are optional)

https://www.youtube.com/watch?v=_Y-F5k3lIDQ
https://www.youtube.com/watch?v=GUlaqYumsGk
https://www.youtube.com/watch?v=zvwrO_Idbvo