JSX Template Formatting
Finally feeling like I like this formatting for things. Was unsure about it earlier. I'm going to continue with using more spacing where possible, I think that's a big help with helping ensure better readability.

I thought the SYL song "Skeksis" was called "Skelesis" this whole time, was reading it wrong. Looks like it's from "The Dark Crystal", which I heard from the DT Podcast it was one of his favorite shows growing up. So I'm going to use "Skelesis" as a name for one of my songs. It can be a little nod to SYL :)

https://www.youtube.com/watch?v=agrAUjogAvo (This was a fun listen haha, curious about looking into the rest of the album now haha)
https://www.youtube.com/watch?v=NEKHOzrasdg
https://www.youtube.com/watch?v=6toI7jk-Prk