Simple General Formatting
Cleaning the few little things up since having reworked a bunch. Intended to leave this to the end of things to help prevent as many breaking changes when trying to diff all of the things I modified over the last while.

These are the last few commits of the `exp-solid` branch, now I'm back on the `dev` branch and going to be pushing the reworks up to `main` next.

Field Music - How Should I Know If You've Changed?