OpenGraph Meta Tags (Partial)
Still need a proper README for the project, as well as demo screenshots, and an OpenGraph image to render for link sharing. Got these from STE, nice and simple to add these few extra nice things.

I didn't add these back when I made the project, I wanted to leave it until a later point when the project was more established (now :D).

https://www.opengraph.xyz/ (Great tool, have used before for adding these to STE as well!)