Solid-CodeMirror
CodeMirror still seems to be a big heck of a hassle to set up, what the heck? I wish it was just simple to put together. The styling for it just doesn't respect the CSS I'm trying to apply to it, and just aagh. It's most certainly me that's the issue in setting it up, it just seems to never go quite right when I set it up.

This is the setup I used just for this demo.
https://github.com/nimeshnayaju/solid-codemirror
https://github.com/Team-Lodestone/codemirror-lang-snbt (I wanted to set this up as well, but it doesn't use regular JS files for the syntax spec, so it needs an additional build step just to work. That's the kind of thing that bugs me about this. Why not use regular language-supported things? Once again, I'm likely the one confused here)

Did some deep dive searching about implementing CodeMirror a few weeks ago, been meaning to document all of the articles I found for it.

https://davidmyers.dev/blog/how-to-build-a-code-editor-with-codemirror-6-and-typescript/introduction
https://www.adamcollier.co.uk/posts/adding-codemirror-6-to-a-react-project
https://codemirror.net/examples/styling/#themes
https://www.bayanbennett.com/posts/failing-to-add-codemirror-6-and-then-succeeding-devlog-004/
https://www.reddit.com/r/swift/comments/r8206e/is_anyone_working_on_a_port_of_codemirror_6_to/
https://github.com/khoi/codemirror-swift
https://github.com/riccardoperra/solid-codemirror
https://github.com/nimeshnayaju/solid-codemirror
https://solid-codemirror.vercel.app/
https://www.youtube.com/watch?v=juKv1aHEikI
https://www.youtube.com/watch?v=5I4ZkmMS4-0&pp=ygUYemlnIHByb2dyYW1taW5nIGxhbmd1YWdl
https://www.youtube.com/watch?v=zVyfRTckDqA (PDF fillable form possible solution!)

Been trying to record more music lately, even when I don't quite feel like it. The goal is again to try making things to see what they sound like, rather than trying to make a specific thing. It doesn't have to be good, it doesn't have to be bad. It just has to be, I guess.